+++ src/components/index.ts
@@ -8,6 +8,8 @@ export { Footer } from "./Footer";
 export { Header } from "./Header";
 export { NumericInput } from "./NumericInput";
 export { ParentInputCard } from "./ParentInputCard";
+export { PeriodToggle, type PeriodUnit } from "./PeriodToggle";
+export { PeriodNumericField, type PeriodNumericFieldProps } from "./PeriodNumericField";
 export { SettlementBanner } from "./SettlementBanner";
 export { Tooltip } from "./Tooltip";
 export { ChangelogModal } from "./changelog";